refactor(ddui): move Materializer into base UI module
