+++ src-test/common/TestDataDrivenUI.cpp
@@ -24,11 +24,11 @@
 #include "ll/api/ui/base/ScreenSession.h"
 #include "ll/api/ui/form/CustomForm.h"
 #include "ll/api/ui/form/MessageBox.h"
+#include "ll/core/ui/base/Materializer.h"
+#include "ll/core/ui/base/UIRawMessage.h"
 #include "ll/core/ui/form/CustomFormModel.h"
 #include "ll/core/ui/form/FormRuntime.h"
-#include "ll/core/ui/form/FormValue.h"
 #include "ll/core/ui/form/MessageBoxModel.h"
-#include "ll/core/ui/base/UIRawMessage.h"
 
 #include "mc/deps/cereal/schema/dynamic/DynamicValue.h"
 #include "mc/platform/UUID.h"
@@ -856,8 +856,8 @@ TEST(DataDrivenUITest, CustomFormComponentsCoverEveryStableOption) {
     auto             dropdownData = Dropdown{
         "Dropdown",
         dropdown,
-                    {{UIRawMessage::text("Two"), 2.0, UIRawMessage::text("Item description")}},
-                    {.description = "Dropdown description", .disabled = true, .visible = false}
+        {{UIRawMessage::text("Two"), 2.0, UIRawMessage::text("Item description")}},
+        {.description = "Dropdown description", .disabled = true, .visible = false}
     }.serialize();
     EXPECT_EQ(dropdownData["description"].asString(), "Dropdown description");
     EXPECT_TRUE(dropdownData["disabled"].asBool());
@@ -874,7 +874,7 @@ TEST(DataDrivenUITest, CustomFormComponentsCoverEveryStableOption) {
         slider,
         NumberValue{-1.0},
         NumberValue{5.0},
-                    {.description = "Slider description", .disabled = true, .step = 0.5, .visible = false}
+        {.description = "Slider description", .disabled = true, .step = 0.5, .visible = false}
     }.serialize();
     EXPECT_EQ(sliderData["description"].asString(), "Slider description");
     EXPECT_TRUE(sliderData["disabled"].asBool());
@@ -889,7 +889,7 @@ TEST(DataDrivenUITest, CustomFormComponentsCoverEveryStableOption) {
     auto             textFieldData = TextField{
         "Text",
         text,
-                    {.description = "Text description", .disabled = true, .visible = false}
+        {.description = "Text description", .disabled = true, .visible = false}
     }.serialize();
     EXPECT_EQ(textFieldData["description"].asString(), "Text description");
     EXPECT_TRUE(textFieldData["disabled"].asBool());
@@ -899,7 +899,7 @@ TEST(DataDrivenUITest, CustomFormComponentsCoverEveryStableOption) {
     auto              toggleData = Toggle{
         "Toggle",
         toggled,
-                     {.description = "Toggle description", .disabled = true, .visible = false}
+        {.description = "Toggle description", .disabled = true, .visible = false}
     }.serialize();
     EXPECT_EQ(toggleData["description"].asString(), "Toggle description");
     EXPECT_TRUE(toggleData["disabled"].asBool());
